refactor(runtime): split methods_mut.rs into cohesive submodules (§7-8)
Split the 4595-line `src/runtime/methods_mut.rs` into 6 sibling files:

- methods_mut.rs (447 lines): core helpers — hash assignment, rw accessor
  normalization, push/unshift arg normalization, inc/dec, array/hash/mixin
  identity propagation
- methods_mut_dispatch.rs (2239 lines): GIANT — 3 free bit-manipulation fns
  + indivisible `call_method_mut_with_values` (~2184 lines, single method)
- methods_mut_hash.rs (133 lines): hash push/append helpers
- methods_mut_method_lvalue.rs (1111 lines): GIANT — indivisible
  `assign_method_lvalue_with_values` (~1107 lines, single method)
- methods_mut_proxy.rs (148 lines): Proxy callback, FETCH, STORE
- methods_mut_rw_attr.rs (137 lines): rw method attribute target detection
  and indexed attribute assignment
- methods_mut_substr_buf.rs (403 lines): substr-rw, subbuf-rw, buf mutation

Pure relocation — no logic changes. Visibility-only: bumped
`store_qualified_attr`, `normalize_rw_accessor_assignment`,
`normalize_incdec_source_for_mut`, `increment_mut_target_value`,
`decrement_mut_target_value`, `is_buf_like_value` to `pub(crate)`.
