+++ src/runtime/mod.rs
@@ -211,6 +211,12 @@ mod methods_io_dispatch;
 mod methods_match_dispatch;
 mod methods_mixin_dispatch;
 mod methods_mut;
+mod methods_mut_dispatch;
+mod methods_mut_hash;
+mod methods_mut_method_lvalue;
+mod methods_mut_proxy;
+mod methods_mut_rw_attr;
+mod methods_mut_substr_buf;
 mod methods_native_bypass;
 mod methods_object;
 mod methods_pick_roll;
